refactor: moving ui-components/inputs to core/components/inputs
