+++ packages/sanity/src/core/components/inputs/DateInputs/DateTimeInput.tsx
@@ -12,8 +12,8 @@ import {
 } from 'react'
 import FocusLock from 'react-focus-lock'
 
-import {Button} from '../../button'
-import {Popover} from '../../popover'
+import {Button} from '../../../../ui-components/button/Button'
+import {Popover} from '../../../../ui-components/popover/Popover'
 import {type CalendarProps} from './calendar/Calendar'
 import {type CalendarLabels} from './calendar/types'
 import {DatePicker} from './DatePicker'
