+++ packages/sanity/src/core/components/inputs/DateInputs/calendar/Calendar.tsx
@@ -15,10 +15,10 @@ import {
   useRef,
 } from 'react'
 
-import useDialogTimeZone from '../../../../core/scheduledPublishing/hooks/useDialogTimeZone'
-import useTimeZone from '../../../../core/scheduledPublishing/hooks/useTimeZone'
-import {Button} from '../../../button/Button'
-import {TooltipDelayGroupProvider} from '../../../tooltipDelayGroupProvider'
+import {Button} from '../../../../../ui-components/button/Button'
+import {TooltipDelayGroupProvider} from '../../../../../ui-components/tooltipDelayGroupProvider/TooltipDelayGroupProvider'
+import useDialogTimeZone from '../../../../scheduledPublishing/hooks/useDialogTimeZone'
+import useTimeZone from '../../../../scheduledPublishing/hooks/useTimeZone'
 import {CalendarMonth} from './CalendarMonth'
 import {ARROW_KEYS, DEFAULT_TIME_PRESETS, HOURS_24} from './constants'
 import {features} from './features'
