+++ packages/sanity/src/core/components/inputs/DateInputs/calendar/CalendarDay.tsx
