+++ packages/sanity/src/core/form/inputs/DateInputs/CommonDateTimeInput.tsx
@@ -11,8 +11,8 @@ import {
   useState,
 } from 'react'
 
-import {type CalendarLabels} from '../../../../ui-components/inputs/DateInputs/calendar/types'
-import {DateTimeInput} from '../../../../ui-components/inputs/DateInputs/DateTimeInput'
+import {type CalendarLabels} from '../../../components/inputs/DateInputs/calendar/types'
+import {DateTimeInput} from '../../../components/inputs/DateInputs/DateTimeInput'
 import {useTranslation} from '../../../i18n/hooks/useTranslation'
 import {type ParseResult} from './types'
