+++ packages/sanity/src/core/form/inputs/DateInputs/DateInput.tsx
@@ -1,7 +1,7 @@
 import {DEFAULT_DATE_FORMAT, format, parse} from '@sanity/util/legacyDateFormat'
 import {useCallback, useMemo} from 'react'
 
-import {type CalendarLabels} from '../../../../ui-components/inputs/DateInputs/calendar/types'
+import {type CalendarLabels} from '../../../components/inputs/DateInputs/calendar/types'
 import {useTranslation} from '../../../i18n'
 import {set, unset} from '../../patch'
 import {type StringInputProps} from '../../types'
