+++ packages/sanity/src/core/scheduledPublishing/components/dateInputs/base/DateTimeInput.tsx
@@ -20,7 +20,7 @@ import {
 import FocusLock from 'react-focus-lock'
 
 import {Popover} from '../../../../../ui-components'
-import {LazyTextInput} from '../../../../../ui-components/inputs/DateInputs/LazyTextInput'
+import {LazyTextInput} from '../../../../components/inputs/DateInputs/LazyTextInput'
 import {DatePicker} from './DatePicker'
 
 type Props = {
