+++ packages/sanity/src/core/tasks/components/form/fields/DateEditFormField.tsx
@@ -6,8 +6,8 @@ import {type KeyboardEvent, useCallback, useMemo, useRef, useState} from 'react'
 import ReactFocusLock from 'react-focus-lock'
 
 import {Button, Popover} from '../../../../../ui-components'
-import {type CalendarLabels} from '../../../../../ui-components/inputs/DateInputs/calendar/types'
-import {DatePicker} from '../../../../../ui-components/inputs/DateInputs/DatePicker'
+import {type CalendarLabels} from '../../../../components/inputs/DateInputs/calendar/types'
+import {DatePicker} from '../../../../components/inputs/DateInputs/DatePicker'
 import {type FormPatch, type PatchEvent, set, unset} from '../../../../form'
 import {getCalendarLabels} from '../../../../form/inputs/DateInputs/utils'
 import {useDateTimeFormat} from '../../../../hooks'
